+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/bot/ktor/middlewares/TelegramBotMiddlewareBuilder.kt
@@ -1,5 +1,6 @@
 package dev.inmo.tgbotapi.bot.ktor.middlewares
 
+import com.benasher44.uuid.uuid4
 import dev.inmo.micro_utils.common.Warning
 import dev.inmo.tgbotapi.bot.ktor.KtorCallFactory
 import dev.inmo.tgbotapi.bot.ktor.TelegramBotPipelinesHandler
@@ -11,9 +12,10 @@ class TelegramBotMiddlewareBuilder {
     var onBeforeSearchCallFactory: (suspend (request: Request<*>, callsFactories: List<KtorCallFactory>) -> Unit)? = null
     var onBeforeCallFactoryMakeCall: (suspend (request: Request<*>, potentialFactory: KtorCallFactory) -> Unit)? = null
     var onAfterCallFactoryMakeCall: (suspend (result: Any?, request: Request<*>, potentialFactory: KtorCallFactory) -> Any?)? = null
-    var onRequestResultPresented: (suspend (result: Any?, request: Request<*>, resultCallFactory: KtorCallFactory, callsFactories: List<KtorCallFactory>) -> Any?)? = null
+    var onRequestResultPresented: (suspend (result: Any, request: Request<*>, resultCallFactory: KtorCallFactory, callsFactories: List<KtorCallFactory>) -> Any?)? = null
     var onRequestResultAbsent: (suspend (request: Request<*>, callsFactories: List<KtorCallFactory>) -> Any?)? = null
     var onRequestReturnResult: (suspend (result: Result<*>, request: Request<*>, callsFactories: List<KtorCallFactory>) -> Result<Any?>?)? = null
+    var id: String = uuid4().toString()
 
     /**
      * Useful way to set [onRequestException]
@@ -42,7 +44,7 @@ class TelegramBotMiddlewareBuilder {
     /**
      * Useful way to set [onRequestResultPresented]
      */
-    fun doOnRequestResultPresented(block: suspend (result: Any?, request: Request<*>, resultCallFactory: KtorCallFactory, callsFactories: List<KtorCallFactory>) -> Any?) {
+    fun doOnRequestResultPresented(block: suspend (result: Any, request: Request<*>, resultCallFactory: KtorCallFactory, callsFactories: List<KtorCallFactory>) -> Any?) {
         onRequestResultPresented = block
     }
     /**
@@ -67,7 +69,8 @@ class TelegramBotMiddlewareBuilder {
             onAfterCallFactoryMakeCall = onAfterCallFactoryMakeCall,
             onRequestResultPresented = onRequestResultPresented,
             onRequestResultAbsent = onRequestResultAbsent,
-            onRequestReturnResult = onRequestReturnResult
+            onRequestReturnResult = onRequestReturnResult,
+            id = id
         )
     }
 
@@ -82,6 +85,7 @@ class TelegramBotMiddlewareBuilder {
                 onRequestResultPresented = middleware.onRequestResultPresented
                 onRequestResultAbsent = middleware.onRequestResultAbsent
                 onRequestReturnResult = middleware.onRequestReturnResult
+                id = middleware.id
                 additionalSetup()
             }.build()
         }